+++ Elements/src/Geometry/Solids/Sweep.cs
@@ -6,7 +6,7 @@ namespace Elements.Geometry.Solids
     public partial class Sweep : SolidOperation, System.ComponentModel.INotifyPropertyChanged
     {
         private Profile _profile;
-        private Curve _curve;
+        private BoundedCurve _curve;
         private double _startSetback;
         private double _endSetback;
         private double _profileRotation;
@@ -21,7 +21,7 @@ namespace Elements.Geometry.Solids
         /// <param name="profileRotation"></param>
         /// <param name="isVoid"></param>
         [JsonConstructor]
-        public Sweep(Profile @profile, Curve @curve, double @startSetback, double @endSetback, double @profileRotation, bool @isVoid)
+        public Sweep(Profile @profile, BoundedCurve @curve, double @startSetback, double @endSetback, double @profileRotation, bool @isVoid)
             : base(isVoid)
         {
             this._profile = @profile;
@@ -51,7 +51,7 @@ namespace Elements.Geometry.Solids
 
         /// <summary>The curve along which the profile will be swept.</summary>
         [JsonProperty("Curve", Required = Required.AllowNull)]
-        public Curve Curve
+        public BoundedCurve Curve
         {
             get { return _curve; }
             set